Refactor Parser classes
Add BaseParser and refactor existing parsers to inherit from it
