+++ app/services/ccms/client_search_parser.rb
@@ -1,14 +1,7 @@
 module CCMS
-  class ClientSearchParser
+  class ClientSearchParser < BaseParser
     TRANSACTION_ID_PATH = '//Body//ClientInqRS//HeaderRS//TransactionID'.freeze
     RECORD_COUNT_PATH = '//Body//ClientInqRS//RecordCount//RecordsFetched'.freeze
-
-    def initialize(tx_request_id, response)
-      @transaction_request_id = tx_request_id
-      @response = response
-      @doc = Nokogiri::XML(@response)
-      @doc.remove_namespaces!
-    end
 
     def parse
       raise 'Invalid transaction request id' if extracted_transaction_request_id != @transaction_request_id
