+++ app/services/ccms/create_client_parser.rb
@@ -1,14 +1,7 @@
 module CCMS
-  class CreateClientParser
+  class CreateClientParser < BaseParser
     TRANSACTION_ID_PATH = '//Body//ClientAddRS//HeaderRS//TransactionID'.freeze
     STATUS_PATH = '//Body//ClientAddRS//HeaderRS//Status//Status'.freeze
-
-    def initialize(tx_request_id, response)
-      @transaction_request_id = tx_request_id
-      @response = response
-      @doc = Nokogiri::XML(@response)
-      @doc.remove_namespaces!
-    end
 
     def parse
       raise 'Invalid transaction request id' if extracted_transaction_request_id != @transaction_request_id
