+++ app/services/ccms/create_client_status_parser.rb
@@ -1,14 +1,7 @@
 module CCMS
-  class CreateClientStatusParser
+  class CreateClientStatusParser < BaseParser
     TRANSACTION_ID_PATH = '//Body//ClientAddUpdtStatusRS//HeaderRS//TransactionID'.freeze
     STATUS_FREE_TEXT_PATH = '//Body//ClientAddUpdtStatusRS//HeaderRS//Status//StatusFreeText'.freeze
-
-    def initialize(tx_request_id, response)
-      @transaction_request_id = tx_request_id
-      @response = response
-      @doc = Nokogiri::XML(@response)
-      @doc.remove_namespaces!
-    end
 
     def parse
       raise 'Invalid transaction request id' if extracted_transaction_request_id != @transaction_request_id
